requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
anyio=3.1.0=py37hf985489_0
appnope=0.1.2=py37hf985489_1
argon2-cffi=20.1.0=py37hf967b71_2
argparse=1.4.0=pypi_0
async_generator=1.10=py_0
attrs=21.2.0=pyhd8ed1ab_0
babel=2.9.1=pyh44b312d_0
backcall=0.2.0=pyh9f0ad1d_0
backports=1.0=py_2
backports.functools_lru_cache=1.6.4=pyhd8ed1ab_0
blas=1.0=mkl
bleach=3.3.0=pyh44b312d_0
brotlipy=0.7.0=py37hf967b71_1001
bzip2=1.0.8=h1de35cc_0
ca-certificates=2021.5.25=hecd8cb5_1
certifi=2020.12.5=py37hecd8cb5_0
cffi=1.14.5=py37h4d371b4_0
chardet=4.0.0=py37hf985489_1
cryptography=3.4.7=py37hce4a858_0
cycler=0.10.0=py_2
decorator=5.0.9=pyhd8ed1ab_0
defusedxml=0.7.1=pyhd8ed1ab_0
entrypoints=0.3=pyhd8ed1ab_1003
ffmpeg=4.3=h0a44026_0
freetype=2.10.4=ha233b18_0
gettext=0.21.0=h7535e17_0
gmp=6.2.1=h23ab428_2
gnutls=3.6.15=hed9c0bf_0
icu=58.2=h0a44026_3
idna=2.10=pyh9f0ad1d_0
importlib-metadata=4.0.1=py37hf985489_0
intel-openmp=2021.2.0=hecd8cb5_564
ipykernel=5.5.5=py37h85f7c60_0
ipython=7.23.1=py37h85f7c60_0
ipython_genutils=0.2.0=py_1
jedi=0.18.0=py37hf985489_2
jinja2=3.0.1=pyhd8ed1ab_0
jpeg=9b=he5867d9_2
json5=0.9.5=pyh9f0ad1d_0
jsonschema=3.2.0=pyhd8ed1ab_3
jupyter_client=6.1.12=pyhd8ed1ab_0
jupyter_core=4.7.1=py37hf985489_0
jupyter_server=1.8.0=pyhd8ed1ab_0
jupyterlab=3.0.16=pyhd8ed1ab_0
jupyterlab_pygments=0.1.2=pyh9f0ad1d_0
jupyterlab_server=2.5.2=pyhd8ed1ab_0
kiwisolver=1.3.1=py37h23ab428_0
lame=3.100=h1de35cc_0
lcms2=2.12=hf1fd2bf_0
libcxx=10.0.0=1
libffi=3.3=hb1e8313_2
libgfortran=3.0.1=h93005f0_2
libiconv=1.16=h1de35cc_0
libidn2=2.3.1=h9ed2024_0
libpng=1.6.37=ha441bb4_0
libsodium=1.0.18=hbcb3906_1
libtasn1=4.16.0=h9ed2024_0
libtiff=4.1.0=hcb84e12_1
libunistring=0.9.10=h9ed2024_0
libuv=1.40.0=haf1e3a3_0
libxml2=2.9.10=h7cdb67c_3
llvm-openmp=10.0.0=h28b9765_0
lz4-c=1.9.3=h23ab428_0
markupsafe=2.0.1=py37h271585c_0
matplotlib=3.2.2=1
matplotlib-base=3.2.2=py37hddda452_1
matplotlib-inline=0.1.2=pyhd8ed1ab_2
mistune=0.8.4=py37hf967b71_1003
mkl=2021.2.0=hecd8cb5_269
mkl-service=2.3.0=py37h9ed2024_1
mkl_fft=1.3.0=py37h4a7008c_2
mkl_random=1.2.1=py37hb2f4e1b_2
nbclassic=0.2.8=pyhd8ed1ab_0
nbclient=0.5.3=pyhd8ed1ab_0
nbconvert=6.0.7=py37hf985489_3
nbformat=5.1.3=pyhd8ed1ab_0
ncurses=6.2=h0a44026_1
nest-asyncio=1.5.1=pyhd8ed1ab_0
nettle=3.7.2=h230ac6f_1
ninja=1.10.2=hf7b0b51_1
notebook=6.4.0=pyha770c72_0
numpy=1.20.2=py37h4b4dc7a_0
numpy-base=1.20.2=py37he0bd621_0
olefile=0.46=py37_0
openh264=2.1.0=hd9629dc_0
openssl=1.1.1k=h9ed2024_0
packaging=20.9=pyh44b312d_0
pandas=1.2.4=pypi_0
pandoc=2.13=h0d85af4_0
pandocfilters=1.4.2=py_1
parso=0.8.2=pyhd8ed1ab_0
pexpect=4.8.0=pyh9f0ad1d_2
pickleshare=0.7.5=py_1003
pillow=8.2.0=py37h5270095_0
pip=21.1.1=py37hecd8cb5_0
prometheus_client=0.10.1=pyhd8ed1ab_0
prompt-toolkit=3.0.18=pyha770c72_0
ptyprocess=0.7.0=pyhd3deb0d_0
pycparser=2.20=pyh9f0ad1d_2
pygments=2.9.0=pyhd8ed1ab_0
pyopenssl=20.0.1=pyhd8ed1ab_0
pyparsing=2.4.7=pyh9f0ad1d_0
pyrsistent=0.17.3=py37hf967b71_2
pysocks=1.7.1=py37hf985489_3
python=3.7.10=h88f2d9e_0
python-dateutil=2.8.1=py_0
python_abi=3.7=1_cp37m
pytorch=1.8.1=py3.7_0
pytz=2021.1=pyhd8ed1ab_0
pyzmq=20.0.0=py37h23ab428_1
readline=8.1=h9ed2024_0
requests=2.25.1=pyhd3deb0d_0
scipy=1.6.2=py37hd5f7400_1
send2trash=1.5.0=py_0
setuptools=52.0.0=py37hecd8cb5_0
six=1.15.0=py37hecd8cb5_0
sniffio=1.2.0=py37hf985489_1
sqlite=3.35.4=hce871da_0
terminado=0.10.0=py37hf985489_0
testpath=0.5.0=pyhd8ed1ab_0
tk=8.6.10=hb0a8c7a_0
torchaudio=0.8.1=py37
torchvision=0.9.1=py37_cpu
tornado=6.1=py37hf967b71_1
tqdm=4.61.0=pypi_0
traitlets=5.0.5=py_0
typing_extensions=3.7.4.3=pyha847dfd_0
urllib3=1.26.4=pyhd8ed1ab_0
wcwidth=0.2.5=pyh9f0ad1d_2
webencodings=0.5.1=py_1
websocket-client=0.57.0=py37hf985489_4
wheel=0.36.2=pyhd3eb1b0_0
xz=5.2.5=h1de35cc_0
zeromq=4.3.4=h23ab428_0
zipp=3.4.1=pyhd8ed1ab_0
zlib=1.2.11=h1de35cc_3
zstd=1.4.9=h322a384_0
